+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/ResourceExtensions.cs
@@ -14,8 +14,6 @@
 // limitations under the License.
 // </copyright>
 
-using System.Collections.Generic;
-using System.Linq;
 using OpenTelemetry.Resources;
 using OtlpCommon = OpenTelemetry.Proto.Common.V1;
 using OtlpResource = OpenTelemetry.Proto.Resource.V1;