+++ src/OpenTelemetry.Shims.OpenTracing/SpanContextShim.cs
@@ -14,8 +14,6 @@
 // limitations under the License.
 // </copyright>
 
-using System;
-using System.Collections.Generic;
 using OpenTracing;
 
 namespace OpenTelemetry.Shims.OpenTracing