+++ src/OpenTelemetry/Metrics/AggregatorStore.cs
@@ -14,11 +14,8 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using System.Threading;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Metrics