+++ src/OpenTelemetry/Metrics/Builder/MeterProviderBuilderSdk.cs
@@ -16,29 +16,50 @@
 
 #nullable enable
 
+using System.Diagnostics;
+using System.Diagnostics.Metrics;
 using System.Text.RegularExpressions;
 using Microsoft.Extensions.DependencyInjection;
+using OpenTelemetry.Internal;
+using OpenTelemetry.Resources;
 
 namespace OpenTelemetry.Metrics
 {
-    internal sealed class MeterProviderBuilderSdk : MeterProviderBuilderBase
+    /// <summary>
+    /// Stores state used to build a <see cref="MeterProvider"/>.
+    /// </summary>
+    internal sealed class MeterProviderBuilderSdk : MeterProviderBuilder, IMeterProviderBuilder
     {
+        public const int MaxMetricsDefault = 1000;
+        public const int MaxMetricPointsPerMetricDefault = 2000;
+        private const string DefaultInstrumentationVersion = "1.0.0.0";
+
         private static readonly Regex InstrumentNameRegex = new(
             @"^[a-z][a-z0-9-._]{0,62}$", RegexOptions.IgnoreCase | RegexOptions.Compiled);
 
-        public MeterProviderBuilderSdk()
+        private readonly IServiceProvider serviceProvider;
+        private MeterProviderSdk? meterProvider;
+
+        public MeterProviderBuilderSdk(IServiceProvider serviceProvider)
         {
+            this.serviceProvider = serviceProvider;
         }
 
-        public MeterProviderBuilderSdk(IServiceCollection services)
-            : base(services)
-        {
-        }
+        public List<InstrumentationRegistration> Instrumentation { get; } = new();
 
-        public MeterProviderBuilderSdk(MeterProviderBuilderState state)
-            : base(state)
-        {
-        }
+        public ResourceBuilder? ResourceBuilder { get; private set; }
+
+        public MeterProvider? Provider => this.meterProvider;
+
+        public List<MetricReader> Readers { get; } = new();
+
+        public List<string> MeterSources { get; } = new();
+
+        public List<Func<Instrument, MetricStreamConfiguration?>> ViewConfigs { get; } = new();
+
+        public int MaxMetricStreams { get; private set; } = MaxMetricsDefault;
+
+        public int MaxMetricPointsPerMetricStream { get; private set; } = MaxMetricPointsPerMetricDefault;
 
         /// <summary>
         /// Returns whether the given instrument name is valid according to the specification.
@@ -71,6 +92,144 @@ namespace OpenTelemetry.Metrics
             }
 
             return InstrumentNameRegex.IsMatch(customViewName);
+        }
+
+        public void RegisterProvider(MeterProviderSdk meterProvider)
+        {
+            Debug.Assert(meterProvider != null, "meterProvider was null");
+
+            if (this.meterProvider != null)
+            {
+                throw new NotSupportedException("MeterProvider cannot be accessed while build is executing.");
+            }
+
+            this.meterProvider = meterProvider;
+        }
+
+        public override MeterProviderBuilder AddInstrumentation<TInstrumentation>(
+            Func<TInstrumentation> instrumentationFactory)
+        {
+            Debug.Assert(instrumentationFactory != null, "instrumentationFactory was null");
+
+            return this.AddInstrumentation(
+                typeof(TInstrumentation).Name,
+                typeof(TInstrumentation).Assembly.GetName().Version?.ToString() ?? DefaultInstrumentationVersion,
+                instrumentationFactory!());
+        }
+
+        public MeterProviderBuilder AddInstrumentation(
+            string instrumentationName,
+            string instrumentationVersion,
+            object instrumentation)
+        {
+            Debug.Assert(!string.IsNullOrWhiteSpace(instrumentationName), "instrumentationName was null or whitespace");
+            Debug.Assert(!string.IsNullOrWhiteSpace(instrumentationVersion), "instrumentationVersion was null or whitespace");
+            Debug.Assert(instrumentation != null, "instrumentation was null");
+
+            this.Instrumentation.Add(
+                new InstrumentationRegistration(
+                    instrumentationName,
+                    instrumentationVersion,
+                    instrumentation!));
+
+            return this;
+        }
+
+        public MeterProviderBuilder ConfigureResource(Action<ResourceBuilder> configure)
+        {
+            Debug.Assert(configure != null, "configure was null");
+
+            var resourceBuilder = this.ResourceBuilder ??= ResourceBuilder.CreateDefault();
+
+            configure!(resourceBuilder);
+
+            return this;
+        }
+
+        public MeterProviderBuilder SetResourceBuilder(ResourceBuilder resourceBuilder)
+        {
+            Debug.Assert(resourceBuilder != null, "resourceBuilder was null");
+
+            this.ResourceBuilder = resourceBuilder;
+
+            return this;
+        }
+
+        public override MeterProviderBuilder AddMeter(params string[] names)
+        {
+            Debug.Assert(names != null, "names was null");
+
+            foreach (var name in names!)
+            {
+                Guard.ThrowIfNullOrWhitespace(name);
+
+                this.MeterSources.Add(name);
+            }
+
+            return this;
+        }
+
+        public MeterProviderBuilder AddReader(MetricReader reader)
+        {
+            Debug.Assert(reader != null, "reader was null");
+
+            this.Readers.Add(reader!);
+
+            return this;
+        }
+
+        public MeterProviderBuilder AddView(Func<Instrument, MetricStreamConfiguration?> viewConfig)
+        {
+            Debug.Assert(viewConfig != null, "viewConfig was null");
+
+            this.ViewConfigs.Add(viewConfig!);
+
+            return this;
+        }
+
+        public MeterProviderBuilder SetMaxMetricStreams(int maxMetricStreams)
+        {
+            this.MaxMetricStreams = maxMetricStreams;
+
+            return this;
+        }
+
+        public MeterProviderBuilder SetMaxMetricPointsPerMetricStream(int maxMetricPointsPerMetricStream)
+        {
+            this.MaxMetricPointsPerMetricStream = maxMetricPointsPerMetricStream;
+
+            return this;
+        }
+
+        public MeterProviderBuilder ConfigureBuilder(Action<IServiceProvider, MeterProviderBuilder> configure)
+        {
+            Debug.Assert(configure != null, "configure was null");
+
+            configure!(this.serviceProvider, this);
+
+            return this;
+        }
+
+        public MeterProviderBuilder ConfigureServices(Action<IServiceCollection> configure)
+        {
+            throw new NotSupportedException("Services cannot be configured after ServiceProvider has been created.");
+        }
+
+        MeterProviderBuilder IDeferredMeterProviderBuilder.Configure(Action<IServiceProvider, MeterProviderBuilder> configure)
+            => this.ConfigureBuilder(configure);
+
+        internal readonly struct InstrumentationRegistration
+        {
+            public readonly string Name;
+            public readonly string Version;
+            public readonly object Instance;
+
+            internal InstrumentationRegistration(string name, string version, object instance)
+            {
+                this.Name = name;
+                this.Version = version;
+                this.Instance = instance;
+            }
         }
     }
 }