+++ src/OpenTelemetry/Metrics/Metric.cs
@@ -14,8 +14,6 @@
 // limitations under the License.
 // </copyright>
 
-using System;
-using System.Collections.Generic;
 using System.Diagnostics.Metrics;
 
 namespace OpenTelemetry.Metrics