+++ src/OpenTelemetry/Trace/SamplingParameters.cs
@@ -16,7 +16,6 @@
 
 #nullable enable
 
-using System.Collections.Generic;
 using System.Diagnostics;
 
 namespace OpenTelemetry.Trace