+++ src/OpenTelemetry/Trace/SamplingResult.cs
@@ -16,10 +16,6 @@
 
 #nullable enable
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
 namespace OpenTelemetry.Trace
 {
     /// <summary>