+++ test/OpenTelemetry.Instrumentation.AspNetCore.Tests/BasicTests.cs
@@ -655,9 +655,12 @@ namespace OpenTelemetry.Instrumentation.AspNetCore.Tests
                     builder.ConfigureTestServices((IServiceCollection services) =>
                     {
                         services.AddSingleton<ActivityMiddleware.ActivityMiddlewareImpl>(new TestNullHostActivityMiddlewareImpl(activitySourceName, activityName));
-                        services.AddOpenTelemetryTracing((builder) => builder.AddAspNetCoreInstrumentation()
-                        .AddSource(activitySourceName)
-                        .AddInMemoryExporter(exportedItems));
+                        services.AddOpenTelemetry()
+                            .WithTracing(builder => builder
+                                .AddAspNetCoreInstrumentation()
+                                .AddSource(activitySourceName)
+                                .AddInMemoryExporter(exportedItems))
+                            .StartWithHost();
                     });
                     builder.ConfigureLogging(loggingBuilder => loggingBuilder.ClearProviders());
                 })
@@ -692,11 +695,11 @@ namespace OpenTelemetry.Instrumentation.AspNetCore.Tests
             var exportedItems = new List<Activity>();
             void ConfigureTestServices(IServiceCollection services)
             {
-                services.AddOpenTelemetryTracing(options =>
-                {
-                    options.AddAspNetCoreInstrumentation()
-                    .AddInMemoryExporter(exportedItems);
-                });
+                services.AddOpenTelemetry()
+                    .WithTracing(builder => builder
+                        .AddAspNetCoreInstrumentation()
+                        .AddInMemoryExporter(exportedItems))
+                    .StartWithHost();
 
                 // Register ActivitySource here so that it will be used
                 // by ASP.NET Core to create activities